Fix warnings throughout codebase.
-Wunused: Remove unused functions, parameters, and variables.
-Wtype-limits: Fix comparison of int to LONG_MIN and LONG_MAX.
-Wwrite-strings: Remove -Wno-write-strings from (c)make.

+++ classes/DelphesStream.cc
@@ -90,12 +90,12 @@ bool DelphesStream::ReadInt(int &value)
   value = strtol(start, &fBuffer, 10);
   if(errno == ERANGE)
   {
-    if(fFirstLongMin && value == LONG_MIN)
+    if(fFirstLongMin && value == INT_MIN)
     {
       fFirstLongMin = false;
       cout << "** WARNING: too large positive value, return " << value << endl;
     }
-    else if(fFirstLongMax && value == LONG_MAX)
+    else if(fFirstLongMax && value == INT_MAX)
     {
       fFirstLongMax = false;
       cout << "** WARNING: too large negative value, return " << value << endl;
